Renamed enums package
