+++ control-system/src/main/java/com/group9/cleansweep/enums/TileTypeEnum.java
@@ -1,4 +1,4 @@
-package com.group9.cleansweep.Enum;
+package com.group9.cleansweep.enums;
 
 public enum TileTypeEnum {
 	UNKNOWN, OBSTACLE, POWERSTATION
